# frontend/openapi.json
{"openapi": "3.1.0", "info": {"title": "Holywater micro-scene", "version": "0.1.0"}, "paths": {"/api/runs": {"post": {"tags": ["runs"], "summary": "Create Run", "operationId": "create_run_api_runs_post", "requestBody": {"required": true, "content": {"application/json": {"schema": {"$ref": "#/components/schemas/CreateRunRequest"}}}}, "responses": {"201": {"description": "Successful Response", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/RunSchema"}}}}, "422": {"description": "Validation Error", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/HTTPValidationError"}}}}}}, "get": {"tags": ["runs"], "summary": "List Runs", "operationId": "list_runs_api_runs_get", "parameters": [{"name": "limit", "in": "query", "required": false, "schema": {"type": "integer", "maximum": 100, "minimum": 1, "default": 20, "title": "Limit"}}], "responses": {"200": {"description": "Successful Response", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/RunListResponse"}}}}, "422": {"description": "Validation Error", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/HTTPValidationError"}}}}}}}, "/api/runs/{run_id}": {"get": {"tags": ["runs"], "summary": "Get Run", "operationId": "get_run_api_runs__run_id__get", "parameters": [{"name": "run_id", "in": "path", "required": true, "schema": {"type": "string", "title": "Run Id"}}], "responses": {"200": {"description": "Successful Response", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/RunSchema"}}}}, "422": {"description": "Validation Error", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/HTTPValidationError"}}}}}}}, "/api/runs/{run_id}/lines": {"get": {"tags": ["runs"], "summary": "Get Run Lines", "operationId": "get_run_lines_api_runs__run_id__lines_get", "parameters": [{"name": "run_id", "in": "path", "required": true, "schema": {"type": "string", "title": "Run Id"}}], "responses": {"200": {"description": "Successful Response", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/RunLinesResponse"}}}}, "422": {"description": "Validation Error", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/HTTPValidationError"}}}}}}}, "/api/runs/{run_id}/video": {"get": {"tags": ["runs"], "summary": "Get Run Video", "operationId": "get_run_video_api_runs__run_id__video_get", "parameters": [{"name": "run_id", "in": "path", "required": true, "schema": {"type": "string", "title": "Run Id"}}], "responses": {"200": {"description": "Successful Response"}, "422": {"description": "Validation Error", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/HTTPValidationError"}}}}}}}, "/api/runs/{run_id}/review": {"post": {"tags": ["runs"], "summary": "Review Run", "operationId": "review_run_api_runs__run_id__review_post", "parameters": [{"name": "run_id", "in": "path", "required": true, "schema": {"type": "string", "title": "Run Id"}}], "requestBody": {"required": true, "content": {"application/json": {"schema": {"$ref": "#/components/schemas/ReviewDecisionRequest"}}}}, "responses": {"200": {"description": "Successful Response", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/RunSchema"}}}}, "422": {"description": "Validation Error", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/HTTPValidationError"}}}}}}}, "/api/runs/{run_id}/continue": {"post": {"tags": ["runs"], "summary": "Continue Run", "operationId": "continue_run_api_runs__run_id__continue_post", "parameters": [{"name": "run_id", "in": "path", "required": true, "schema": {"type": "string", "title": "Run Id"}}], "responses": {"200": {"description": "Successful Response", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/RunSchema"}}}}, "422": {"description": "Validation Error", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/HTTPValidationError"}}}}}}}, "/api/health": {"get": {"tags": ["health"], "summary": "Health", "operationId": "health_api_health_get", "responses": {"200": {"description": "Successful Response", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/HealthResponse"}}}}}}}}, "components": {"schemas": {"AlignedWordSchema": {"properties": {"op": {"anyOf": [{"$ref": "#/components/schemas/DiffOp"}, {"type": "null"}]}, "expected": {"anyOf": [{"type": "string"}, {"type": "null"}], "title": "Expected"}, "heard": {"anyOf": [{"type": "string"}, {"type": "null"}], "title": "Heard"}}, "type": "object", "required": ["op", "expected", "heard"], "title": "AlignedWordSchema"}, "CreateRunRequest": {"properties": {"idea": {"type": "string", "maxLength": 600, "minLength": 10, "title": "Idea"}}, "type": "object", "required": ["idea"], "title": "CreateRunRequest"}, "DiffOp": {"type": "string", "enum": ["sub", "del", "ins"], "title": "DiffOp"}, "HTTPValidationError": {"properties": {"detail": {"items": {"$ref": "#/components/schemas/ValidationError"}, "type": "array", "title": "Detail"}}, "type": "object", "title": "HTTPValidationError"}, "HealthResponse": {"properties": {"status": {"type": "string", "title": "Status"}}, "type": "object", "required": ["status"], "title": "HealthResponse"}, "LineReportSchema": {"properties": {"shot_id": {"type": "string", "title": "Shot Id"}, "speaker": {"type": "string", "title": "Speaker"}, "speaker_name": {"type": "string", "title": "Speaker Name"}, "text": {"type": "string", "title": "Text"}, "locked": {"type": "boolean", "title": "Locked"}, "attempt": {"anyOf": [{"type": "integer"}, {"type": "null"}], "title": "Attempt"}, "heard": {"anyOf": [{"type": "string"}, {"type": "null"}], "title": "Heard"}, "wer": {"anyOf": [{"type": "number"}, {"type": "null"}], "title": "Wer"}, "diff": {"items": {"$ref": "#/components/schemas/WordDiffSchema"}, "type": "array", "title": "Diff"}, "alignment": {"items": {"$ref": "#/components/schemas/AlignedWordSchema"}, "type": "array", "title": "Alignment"}, "speaker_matches": {"anyOf": [{"type": "boolean"}, {"type": "null"}], "title": "Speaker Matches"}}, "type": "object", "required": ["shot_id", "speaker", "speaker_name", "text", "locked", "attempt", "heard", "wer", "diff", "alignment", "speaker_matches"], "title": "LineReportSchema"}, "ReviewDecisionRequest": {"properties": {"approve": {"type": "boolean", "title": "Approve"}}, "type": "object", "required": ["approve"], "title": "ReviewDecisionRequest"}, "RunLinesResponse": {"properties": {"items": {"items": {"$ref": "#/components/schemas/LineReportSchema"}, "type": "array", "title": "Items"}}, "type": "object", "required": ["items"], "title": "RunLinesResponse"}, "RunListResponse": {"properties": {"items": {"items": {"$ref": "#/components/schemas/RunSummarySchema"}, "type": "array", "title": "Items"}}, "type": "object", "required": ["items"], "title": "RunListResponse"}, "RunSchema": {"properties": {"id": {"type": "string", "title": "Id"}, "idea": {"type": "string", "title": "Idea"}, "status": {"$ref": "#/components/schemas/RunStatus"}, "cost_usd": {"type": "number", "title": "Cost Usd"}, "created_at": {"type": "string", "format": "date-time", "title": "Created At"}, "video_url": {"anyOf": [{"type": "string"}, {"type": "null"}], "title": "Video Url"}, "current_step": {"anyOf": [{"$ref": "#/components/schemas/StepName"}, {"type": "null"}]}, "review_reason": {"anyOf": [{"type": "string"}, {"type": "null"}], "title": "Review Reason"}, "error": {"anyOf": [{"type": "string"}, {"type": "null"}], "title": "Error"}, "updated_at": {"type": "string", "format": "date-time", "title": "Updated At"}, "finished_at": {"anyOf": [{"type": "string", "format": "date-time"}, {"type": "null"}], "title": "Finished At"}, "steps": {"items": {"$ref": "#/components/schemas/RunStepSchema"}, "type": "array", "title": "Steps"}, "can_continue": {"type": "boolean", "title": "Can Continue", "default": false}}, "type": "object", "required": ["id", "idea", "status", "cost_usd", "created_at", "current_step", "review_reason", "error", "updated_at", "finished_at", "steps"], "title": "RunSchema"}, "RunStatus": {"type": "string", "enum": ["queued", "running", "done", "needs_review", "failed"], "title": "RunStatus"}, "RunStepSchema": {"properties": {"id": {"type": "string", "title": "Id"}, "name": {"$ref": "#/components/schemas/StepName"}, "status": {"$ref": "#/components/schemas/StepStatus"}, "attempt": {"type": "integer", "title": "Attempt"}, "detail": {"anyOf": [{"type": "string"}, {"type": "null"}], "title": "Detail"}, "cost_usd": {"type": "number", "title": "Cost Usd"}, "created_at": {"type": "string", "format": "date-time", "title": "Created At"}, "finished_at": {"anyOf": [{"type": "string", "format": "date-time"}, {"type": "null"}], "title": "Finished At"}}, "type": "object", "required": ["id", "name", "status", "attempt", "detail", "cost_usd", "created_at", "finished_at"], "title": "RunStepSchema"}, "RunSummarySchema": {"properties": {"id": {"type": "string", "title": "Id"}, "idea": {"type": "string", "title": "Idea"}, "status": {"$ref": "#/components/schemas/RunStatus"}, "cost_usd": {"type": "number", "title": "Cost Usd"}, "created_at": {"type": "string", "format": "date-time", "title": "Created At"}, "video_url": {"anyOf": [{"type": "string"}, {"type": "null"}], "title": "Video Url"}}, "type": "object", "required": ["id", "idea", "status", "cost_usd", "created_at"], "title": "RunSummarySchema"}, "StepName": {"type": "string", "enum": ["brief", "script", "script_check", "cast", "portraits", "portrait_check", "scene", "render", "scene_check", "shot_fix", "final"], "title": "StepName"}, "StepStatus": {"type": "string", "enum": ["running", "ok", "rejected", "failed"], "title": "StepStatus"}, "ValidationError": {"properties": {"loc": {"items": {"anyOf": [{"type": "string"}, {"type": "integer"}]}, "type": "array", "title": "Location"}, "msg": {"type": "string", "title": "Message"}, "type": {"type": "string", "title": "Error Type"}, "input": {"title": "Input"}, "ctx": {"type": "object", "title": "Context"}}, "type": "object", "required": ["loc", "msg", "type"], "title": "ValidationError"}, "WordDiffSchema": {"properties": {"op": {"$ref": "#/components/schemas/DiffOp"}, "expected": {"anyOf": [{"type": "string"}, {"type": "null"}], "title": "Expected"}, "heard": {"anyOf": [{"type": "string"}, {"type": "null"}], "title": "Heard"}}, "type": "object", "required": ["op", "expected", "heard"], "title": "WordDiffSchema"}}}}